+++ Editor/SceneViewController.cs
@@ -162,6 +162,7 @@ public static class SceneViewController
         Time.timeScale = Mathf.Clamp01(Time.timeScale + 0.2f);
         Debug.LogWarning($"[Alt + +]Time.timeScale => {Time.timeScale}");
     }
+    
     #endregion